+++ src/app/home/MissionList.tsx
@@ -1,76 +1,11 @@
 'use client';
 
 import Link from 'next/link';
-import { MissionStatus } from '@/apis/schema/mission';
-import { MissionListSkeleton } from '@/app/home/home.styles';
-import MissionBadge from '@/app/home/MissionBadge';
-import MissionEmptyList from '@/app/home/MissionEmptyList';
+import MissionListInner from '@/app/home/MissionInnerList';
 import Icon from '@/components/Icon';
-import { TwoLineListItem } from '@/components/ListItem';
-import { MISSION_CATEGORY_LABEL } from '@/constants/mission';
 import { ROUTER } from '@/constants/router';
 import { css } from '@/styled-system/css';
 import { flex } from '@/styled-system/patterns';
-
-import { useMissions } from './home.hooks';
-
-function MissionList() {
-  return (
-    <div className={containerCss}>
-      <Header />
-      <ul className={listCss}>
-        {/* TODO : Suspense로 리팩토링 */}
-        {/* TODO : 미션 최근 순 정렬 */}
-        {/* TODO : 완료된 미션은 하단 정렬 */}
-        <MissionListInner />
-      </ul>
-    </div>
-  );
-}
-
-export default MissionList;
-
-export function MissionListInner() {
-  const { missionList, isLoading, progressMissionId } = useMissions();
-
-  if (isLoading) {
-    return <MissionListSkeleton />;
-  }
-
-  if (missionList.length === 0) {
-    return (
-      <div className={containerCss}>
-        <MissionEmptyList />
-      </div>
-    );
-  }
-
-  return (
-    <>
-      {missionList.map((item) => {
-        const isProgressingMission = progressMissionId === String(item.missionId);
-        const status = isProgressingMission ? MissionStatus.PROGRESSING : item.missionStatus;
-
-        const missionId = item.missionId.toString();
-        const moveHref = isProgressingMission
-          ? ROUTER.MISSION.STOP_WATCH(missionId)
-          : item.missionRecordId && status === MissionStatus.REQUIRED
-            ? ROUTER.RECORD.CREATE(item.missionRecordId.toString())
-            : ROUTER.MISSION.DETAIL(missionId);
-        return (
-          <Link href={moveHref} key={item.missionId}>
-            <TwoLineListItem
-              badgeElement={<MissionBadge status={status} />}
-              name={item.name}
-              subName={MISSION_CATEGORY_LABEL[item.category].label}
-              imageUrl={MISSION_CATEGORY_LABEL[item.category].imgUrl}
-            />
-          </Link>
-        );
-      })}
-    </>
-  );
-}
 
 function Header() {
   return (
@@ -82,6 +17,19 @@ function Header() {
     </h2>
   );
 }
+
+function MissionList() {
+  return (
+    <div className={containerCss}>
+      <Header />
+      <ul className={listCss}>
+        <MissionListInner />
+      </ul>
+    </div>
+  );
+}
+
+export default MissionList;
 
 const containerCss = css({
   height: '100%',
